requirements.txt
# requirements.txt
# -----------------
# Python >= 3.9 recommended.
# Install with:  pip install -r requirements.txt

astropy>=5.3
astroquery>=0.4.6
numpy>=1.24
pandas>=2.0
scipy>=1.10
matplotlib>=3.7
PyYAML>=6.0
